nlp_dashboard/requirements.txt
streamlit
transformers
nltk
newspaper3k
newspaper4k
wordcloud
matplotlib
